print int now int32

+++ core/vm/common.cpp
@@ -1950,7 +1950,7 @@ bool TrapProcessor::StdOutInt(StackProgram* program, size_t* inst, size_t* &op_s
 #ifdef _DEBUG
   wcout << L"  STD_OUT_INT" << endl;
 #endif
-  wcout << (long)PopInt(op_stack, stack_pos);
+  wcout << (int32_t)PopInt(op_stack, stack_pos);
 
   return true;
 }
